30thNov-voicebot-workshop/requirements.txt
sounddevice
scipy
whisper
google-generativeai
python-dotenv
gtts
playsound==1.2.2
pyaudio
streamlit
sppechRecognition